Refactor theme handling into separate mixin
